pyproject.toml
[project]
name = "deepagents"
version = "0.0.3"
description = "General purpose 'deep agent' with sub-agent spawning, todo list capabilities, and mock file system. Built on LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=0.2.6",
    "langchain-anthropic>=0.1.23",
    "langchain>=0.2.14",
    "tavily-python",
    "langchain-cerebras",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.287",
    "tavily-python>=0.3.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]

examples = [
    "tavily-python>=0.3.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deepagents"]
[tool.setuptools.package-dir]
"deepagents" = "src/deepagents"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[dependency-groups]
dev = [
    "pytest-cov>=6.2.1",
]
